extract: ErpServiceTestSuites, Test_ErpServiceDouble
